requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
rich>=13.7.0
selenium>=4.15.0
webdriver-manager>=4.0.0
